+++ src/index.ts
@@ -35,14 +35,13 @@ export * from './components/InfiniteList'
 export * from './components/Label'
 export * from './components/Logo'
 export * from './components/Menu'
-export * from './components/MenuSmall'
+export * from './components/Sidebar'
 export * from './components/Page'
 export * from './components/PieGraph'
 export * from './components/Provider'
 export * from './components/RadioButton'
 export * from './components/ResultCard'
 
-export * from './components/Sidebar'
 export * from './components/Steps'
 export * from './components/Select'
 export * from './components/Separator'
